refactor: authentication パッケージのクライアントを用途別ディレクトリに整理
フラットに並んでいた各クライアントを認証手段(oauth/passkey/password/session)・
共通基盤(supabase)・テスト専用(admin)のディレクトリに分割し、用途を構成から読めるようにした。
ディレクトリが文脈を与えるため冗長な接頭辞を落とし、公開 API(index.ts の export)は変更しない。
