+++ application/packages/authentication/src/index.ts
@@ -1,4 +1,4 @@
-export { AuthAdminClient, type AuthAdminConfig, type AuthUserSummary } from './auth-admin-client'
+export { AuthAdminClient, type AuthAdminConfig, type AuthUserSummary } from './admin/client'
 export {
   AuthError,
   InvalidCredentialsError,
@@ -8,9 +8,9 @@ export {
   UnexpectedAuthError,
   UserAlreadyExistsError,
 } from './errors'
-export { OAuthClient } from './oauth-client'
-export { OAUTH_PROVIDERS, type OAuthProvider } from './oauth-provider'
-export { PasskeyClient } from './passkey-client'
-export { PasswordAuthClient } from './password-auth-client'
-export { SessionClient } from './session-client'
-export { type AuthClientConfig, authClientConfig, type AuthRequestContext } from './supabase-client'
+export { OAuthClient } from './oauth/client'
+export { OAUTH_PROVIDERS, type OAuthProvider } from './oauth/provider'
+export { PasskeyClient } from './passkey/client'
+export { PasswordAuthClient } from './password/client'
+export { SessionClient } from './session/client'
+export { type AuthClientConfig, authClientConfig, type AuthRequestContext } from './supabase/client'
