+++ application/packages/authentication/src/password/client.ts
@@ -5,13 +5,13 @@ import {
   InvalidCredentialsError,
   UnexpectedAuthError,
   UserAlreadyExistsError,
-} from './errors'
+} from '../errors'
 import {
   type AuthClientConfig,
   createBackendClient,
   type SupabaseAuthClient,
-} from './supabase-client'
-import { callSupabase } from './supabase-result'
+} from '../supabase/client'
+import { callSupabase } from '../supabase/result'
 
 interface PasswordCredentials {
   email: string
